Now asks user to delete a folder of frames after split is done

+++ src/com/company/Util/DirectoryUtilities.java
@@ -1,24 +1,30 @@
 package com.company.Util;
 
 import java.io.File;
+import java.util.Scanner;
 
 public class DirectoryUtilities {
 
     public static void deleteDirectory(String path) {
-        File directory = new File(path);
-        if (directory.exists()) {
-            File[] files = directory.listFiles();
-            if (null != files) {
-                for (int i = 0; i < files.length; i++) {
-                    if (files[i].isDirectory()) {
-                        deleteDirectory(files[i].getPath());
-                    } else {
-                        files[i].delete();
+        System.out.println("Are you sure you want to delete this folder: " + path + "? y/n");
+        Scanner keyboard = new Scanner(System.in);
+        String answer = keyboard.nextLine();
+        if (answer.equals("y")) {
+            File directory = new File(path);
+            if (directory.exists()) {
+                File[] files = directory.listFiles();
+                if (null != files) {
+                    for (int i = 0; i < files.length; i++) {
+                        if (files[i].isDirectory()) {
+                            deleteDirectory(files[i].getPath());
+                        } else {
+                            files[i].delete();
+                        }
                     }
                 }
             }
+            directory.delete();
         }
-        directory.delete();
     }
 
 }
